Dwi Putra Satria Utama_VIX_IDX Partners/deployment/requirements.txt
streamlit
pandas
seaborn=0.13.2
matplotlib
plotly
pillow
numpy
scikit-learn==1.2.2
imbalanced-learn==0.11.0